+++ src/lib.rs
@@ -1,15 +1,16 @@
 mod admin_websocket;
-mod app_agent_websocket;
 mod app_websocket;
+mod app_websocket_inner;
 mod error;
 mod signing;
 
 pub use admin_websocket::{AdminWebsocket, AuthorizeSigningCredentialsPayload, EnableAppResponse};
-pub use app_agent_websocket::{AppAgentWebsocket, ZomeCallTarget};
-pub use app_websocket::AppWebsocket;
+pub use app_websocket::{AppWebsocket, ZomeCallTarget};
 pub use error::{ConductorApiError, ConductorApiResult};
 pub use holochain_conductor_api::{
-    AdminRequest, AdminResponse, AppInfo, AppRequest, AppResponse, AppStatusFilter, ZomeCall,
+    AdminRequest, AdminResponse, AppAuthenticationRequest, AppAuthenticationToken,
+    AppAuthenticationTokenIssued, AppInfo, AppRequest, AppResponse, AppStatusFilter,
+    IssueAppAuthenticationTokenPayload, ZomeCall,
 };
 pub use holochain_types::{
     app::{InstallAppPayload, InstalledAppId},