+++ src/signing/lair_signing.rs
@@ -6,26 +6,27 @@ use holochain_zome_types::{
     capability::CapSecret, cell::CellId, dependencies::holochain_integrity_types::Signature,
 };
 use lair_keystore_api::LairClient;
+use parking_lot::RwLock;
 use std::collections::HashMap;
 use std::sync::Arc;
 
 pub struct LairAgentSigner {
     lair_client: Arc<LairClient>,
-    credentials: HashMap<CellId, AgentPubKey>,
+    credentials: Arc<RwLock<HashMap<CellId, AgentPubKey>>>,
 }
 
 impl LairAgentSigner {
     pub fn new(lair_client: Arc<LairClient>) -> Self {
         Self {
             lair_client,
-            credentials: HashMap::new(),
+            credentials: Arc::new(RwLock::new(HashMap::new())),
         }
     }
 
     /// Add credentials for a cell to the signer.
     /// The provenance should be the `agent_pub_key` that the cell is running as.
     pub fn add_credentials(&mut self, cell_id: CellId, provenance: AgentPubKey) {
-        self.credentials.insert(cell_id, provenance);
+        self.credentials.write().insert(cell_id, provenance);
     }
 }
 
@@ -48,7 +49,7 @@ impl AgentSigner for LairAgentSigner {
     }
 
     fn get_provenance(&self, cell_id: &CellId) -> Option<AgentPubKey> {
-        self.credentials.get(cell_id).cloned()
+        self.credentials.read().get(cell_id).cloned()
     }
 
     /// Not used with Lair signing. If you have access to Lair then you don't need to prove you