+++ CHANGELOG.md
@@ -5,9 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## \[Unreleased\]
 
 ### Added
+- New admin call `issue_app_auth_token` which allows you to issue an app auth token. This is now required when creating
+  an app websocket connection. See the example for `AppWebsocket::connect` for how to use this.
+- Missing app interface function `list_wasm_host_functions` has been added.
+
 ### Changed
+- **BREAKING**: The admin call `list_app_interfaces` now returns a `Vec<AppInterfaceInfo>` instead of a `Vec<u16>`. You can map the response to a `Vec<u16>` to get the previous result.
+- **BREAKING**: The admin call `attach_app_interface` now takes an additional parameter of type `InstalledAppId` which allows you to restrict an app interface to a single app.
+- **BREAKING**: The app call `app_info` no longer takes an `installed_app_id` parameter. The conductor uses the app that you have authenticated with.
+
 ### Fixed
 ### Removed
+- **BREAKING**: The old `AppWebsocket` is gone, its functionality was merged into the `AppAgentWebsocket` which has been renamed to `AppWebsocket`.
 
 ## 2024-03-27: v0.5.0-dev.31
 ### Changed